More parser and cleaner code
- Move more logic into the parser module where it makes sense
- Remove passing around g/request/session, flask has those as global for a reason
- Change utils -> cache, because that makes sense
- Add a dataclass module and move all dataclasses into it
