+++ app/blueprints/awards/dxcc.py
@@ -1,7 +1,7 @@
-from flask import g, render_template, request, session, url_for
+from flask import render_template, url_for
 
+from ...cache import get_award_details
 from ...urls import DXCC_PAGE_URL
-from ...utils import get_award_details
 from ..auth.wrappers import login_required
 from .base import bp
 
@@ -9,17 +9,12 @@ from .base import bp
 @bp.get("/dxcc")
 @login_required(next_page="awards.dxcc")
 def dxcc():
-    dxcc_details, dxcc_parsed_at = get_award_details(
-        award="dxcc",
-        g=g,
-        request=request,
-        session=session,
-    )
+    dxcc_details, was_parsed_at = get_award_details(award="dxcc")
 
     return render_template(
         "award.html",
         awards=dxcc_details,
-        parsed_at=dxcc_parsed_at.strftime("%d/%m/%Y, %H:%M:%S"),
+        parsed_at=was_parsed_at.strftime("%d/%m/%Y, %H:%M:%S"),
         force_reload=url_for("awards.dxcc", force_reload=True),
         page_url=DXCC_PAGE_URL,
         award_name="DXCC",
