+++ app/database/table_declarations/base.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 from sqlalchemy import MetaData
 from sqlalchemy.orm import DeclarativeBase
