+++ app/dataclasses/__init__.py
@@ -1,0 +1,4 @@
+from .awards_detail import AwardsDetail
+from .qso_detail import QSODetail
+from .row import Row
+from .triple_detail import TripleDetail
