+++ app/dataclasses/awards_detail.py
