+++ app/dataclasses/row.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Row:
+    label: str
+    value: str
