+++ app/dataclasses/triple_detail.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class TripleDetail:
+    op: str
+    state: str
+    cw: str
+    phone: str
+    digital: str
