+++ app/templates/award_details.html
@@ -17,7 +17,7 @@
         WAZ Zone
         {% endif %}
       </th>
-      <th style="text-align: center;">{{th}}</th>
+      <th style="text-align: center;">{{table_header}}</th>
     </tr>
   </thead>
   <tbody>
