+++ app/templates/qsl_details.html
@@ -3,20 +3,27 @@
 {% block content %}
 <table class="table ">
   <tbody>
-    {% for qsl_detail in qsl_details %}
-
-    <tr>
-    <td style="text-align: left; {% if qsl_detail.label == 'Call Sign' %} background-color: #e3f2fd;
-    {% elif qsl_detail.label == 'Worked' %} background-color: #e3f2fd; {% endif %}">            
-    <b>{{ qsl_detail.label }}</b></td>
-    
-    <td style="text-align: left; {% if qsl_detail.label == 'Call Sign' %} background-color: #e3f2fd;
-    {% elif qsl_detail.label == 'Worked' %} background-color: #e3f2fd; {% endif %}">            
-    {{ qsl_detail.value }}</td>
-    </tr>
+    {% for row in qsl_detail.rows %}
+      <tr>
+        <td 
+          style="text-align: left; 
+          {% if row.label == 'Call Sign' or row.label == 'Worked' %} 
+          background-color: #e3f2fd;
+          {% endif %}"
+        >            
+          <b>{{ row.label }}</b>
+        </td>
+        
+        <td 
+          style="text-align: left; 
+          {% if row.label == 'Call Sign' or row.label == 'Worked' %} 
+          background-color: #e3f2fd;
+          {% endif %}"
+        >              
+          {{ row.value }}
+        </td>
+      </tr>
     {% endfor %}
-
-
   </tbody>
 </table>
 {% endblock %}
